genaihindi_2/requirements.txt
openai
langchain
langchain-community
langchain-openai
streamlit
python-dotenv
PyPDF2


-e .